refactor: stories of notice, paginationV2 and passwordCheck
