+++ src/components/Notice/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Notice from '..'
+
+export const Template: ComponentStory<typeof Notice> = args => (
+  <Notice {...args}>This is a notice</Notice>
+)
