+++ src/components/PaginationV2/__stories__/Basic.stories.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template.stories'
+
+export const Basic = Template.bind({})
+Basic.args = {
+  disabled: false,
+  onChange: () => {},
+  page: 3,
+  pageCount: 10,
+}
