+++ src/components/PaginationV2/__stories__/Disabled.stories.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template.stories'
+
+export const Disabled = Template.bind({})
+Disabled.args = {
+  disabled: true,
+  onChange: () => {},
+  page: 3,
+  pageCount: 10,
+}
