+++ src/components/PaginationV2/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Pagination from '..'
+
+export const Template: ComponentStory<typeof Pagination> = args => (
+  <Pagination {...args} />
+)
