+++ src/components/PaginationV2/__stories__/index.stories.tsx
@@ -1,5 +1,3 @@
-import { Story } from '@storybook/react'
-import { ComponentProps } from 'react'
 import Pagination from '..'
 
 export default {
@@ -15,24 +13,6 @@ export default {
   title: 'Components/Navigation/PaginationV2',
 }
 
-const Template: Story<ComponentProps<typeof Pagination>> = args => (
-  <Pagination {...args} />
-)
-
-export const Default = Template.bind({})
-
-export const Basic = Template.bind({})
-Basic.args = {
-  disabled: false,
-  onChange: () => {},
-  page: 3,
-  pageCount: 10,
-}
-
-export const Disabled = Template.bind({})
-Disabled.args = {
-  disabled: true,
-  onChange: () => {},
-  page: 3,
-  pageCount: 10,
-}
+export { Playground } from './Playground.stories'
+export { Basic } from './Basic.stories'
+export { Disabled } from './Disabled.stories'
