+++ src/components/PasswordCheck/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import PasswordCheck from '..'
+
+export const Template: ComponentStory<typeof PasswordCheck> = args => (
+  <PasswordCheck {...args}>Button</PasswordCheck>
+)
